Report-Analyzer/requirements.txt
pandas
tk
pillow
